requirements.txt
# Website Assembler Dependencies

# Core Dependencies
jinja2>=3.1.0
requests>=2.31.0

# Visual Validation
playwright>=1.40.0

# HTML/CSS Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utility
python-dotenv>=1.0.0

# AI Generation
google-generativeai>=0.3.0
